sharing tests
