+++ tests/Wpf/Prism.Container.Wpf.Shared/Fixtures/Bootstrapper/BootstrapperNullModuleManagerFixture.cs
@@ -1,0 +1,18 @@
+﻿using Prism.Container.Wpf.Mocks;
+using Xunit;
+
+namespace Prism.Container.Wpf.Tests.Bootstrapper
+{
+    public class BootstrapperNullModuleManagerFixture
+    {
+        [Fact]
+        public void RunShouldNotCallInitializeModulesWhenModuleManagerNotFound()
+        {
+            var bootstrapper = new NullModuleManagerBootstrapper();
+
+            bootstrapper.Run();
+
+            Assert.False(bootstrapper.InitializeModulesCalled);
+        }
+    }
+}
