+++ tests/Wpf/Prism.Container.Wpf.Shared/Fixtures/Bootstrapper/BootstrapperRunMethodFixture.cs
@@ -1,0 +1,417 @@
+using System;
+using System.Windows;
+using System.Windows.Controls;
+using Moq;
+using Prism.Container.Wpf.Mocks;
+using Prism.Container.Wpf.Tests;
+using Prism.Events;
+using Prism.Ioc;
+using Prism.Logging;
+using Prism.Modularity;
+using Prism.Regions;
+using Xunit;
+
+namespace Prism.Container.Wpf.Tests
+{
+    [Collection(ContainerHelper.CollectionName)]
+    public partial class BootstrapperRunMethodFixture
+    {
+        [StaFact]
+        public void CanRunBootstrapper()
+        {
+            var bootstrapper = new MockBootstrapper();
+            bootstrapper.Run();
+        }
+
+        [StaFact]
+        public void RunShouldNotFailIfReturnedNullShell()
+        {
+            var bootstrapper = new MockBootstrapper { ShellObject = null };
+            bootstrapper.Run();
+        }
+
+        [StaFact]
+        public void RunSetsCurrentContainer()
+        {
+            var bootstrapper = new MockBootstrapper();
+            bootstrapper.Run();
+
+            Assert.NotNull(ContainerLocator.Container);
+            Assert.IsType(ContainerHelper.ContainerExtensionType, ContainerLocator.Container);
+        }
+
+        [StaFact]
+        public void RunShouldInitializeContainer()
+        {
+            var bootstrapper = new MockBootstrapper();
+            var container = bootstrapper.BaseContainer;
+
+            Assert.Null(container);
+
+            bootstrapper.Run();
+
+            container = bootstrapper.BaseContainer;
+
+            Assert.NotNull(container);
+            Assert.IsType(ContainerHelper.BaseContainerType, container);
+        }
+
+        [StaFact]
+        public void RunShouldCallInitializeModules()
+        {
+            var bootstrapper = new MockBootstrapper();
+            bootstrapper.Run();
+
+            Assert.True(bootstrapper.InitializeModulesCalled);
+        }
+
+        [StaFact]
+        public void RunShouldCallConfigureDefaultRegionBehaviors()
+        {
+            var bootstrapper = new MockBootstrapper();
+
+            bootstrapper.Run();
+
+            Assert.True(bootstrapper.ConfigureDefaultRegionBehaviorsCalled);
+        }
+
+        [StaFact]
+        public void RunShouldCallConfigureRegionAdapterMappings()
+        {
+            var bootstrapper = new MockBootstrapper();
+
+            bootstrapper.Run();
+
+            Assert.True(bootstrapper.ConfigureRegionAdapterMappingsCalled);
+        }
+
+        [StaFact]
+        public void RunShouldAssignRegionManagerToReturnedShell()
+        {
+            var bootstrapper = new MockBootstrapper();
+
+            bootstrapper.Run();
+
+            Assert.NotNull(RegionManager.GetRegionManager(bootstrapper.BaseShell));
+        }
+
+        [StaFact]
+        public void RunShouldCallCreateLogger()
+        {
+            var bootstrapper = new MockBootstrapper();
+
+            bootstrapper.Run();
+
+            Assert.True(bootstrapper.CreateLoggerCalled);
+        }
+
+        [StaFact]
+        public void RunShouldCallCreateModuleCatalog()
+        {
+            var bootstrapper = new MockBootstrapper();
+
+            bootstrapper.Run();
+
+            Assert.True(bootstrapper.CreateModuleCatalogCalled);
+        }
+
+        [StaFact]
+        public void RunShouldCallConfigureModuleCatalog()
+        {
+            var bootstrapper = new MockBootstrapper();
+
+            bootstrapper.Run();
+
+            Assert.True(bootstrapper.ConfigureModuleCatalogCalled);
+        }
+
+        [StaFact]
+        public void RunShouldCallCreateContainer()
+        {
+            var bootstrapper = new MockBootstrapper();
+
+            bootstrapper.Run();
+
+            Assert.True(bootstrapper.CreateContainerCalled);
+        }
+
+        [StaFact]
+        public void RunShouldCallCreateShell()
+        {
+            var bootstrapper = new MockBootstrapper();
+
+            bootstrapper.Run();
+
+            Assert.True(bootstrapper.CreateShellCalled);
+        }
+
+        [StaFact]
+        public void RunShouldCallConfigureContainer()
+        {
+            var bootstrapper = new MockBootstrapper();
+
+            bootstrapper.Run();
+
+            Assert.True(bootstrapper.ConfigureContainerCalled);
+        }
+
+        [StaFact]
+        public void SetsContainerLocatorCurrentContainer()
+        {
+            ContainerLocator.SetCurrent(null);
+            Assert.Null(ContainerLocator.Container);
+            var bootstrapper = new MockBootstrapper();
+
+            bootstrapper.Run();
+
+            Assert.NotNull(ContainerLocator.Container);
+            Assert.Same(bootstrapper.Container, ContainerLocator.Container.GetBaseContainer());
+        }
+
+        [StaFact]
+        public void RunShouldCallConfigureViewModelLocator()
+        {
+            var bootstrapper = new MockBootstrapper();
+
+            bootstrapper.Run();
+
+            Assert.True(bootstrapper.ConfigureViewModelLocatorCalled);
+        }
+
+        [StaFact]
+        public void ModuleManagerRunCalled()
+        {
+            // Have to use a non-mocked container because of IsRegistered<> extension method, Registrations property,and ContainerRegistration
+            var mockedModuleInitializer = new Mock<IModuleInitializer>();
+            var mockedModuleManager = new Mock<IModuleManager>();
+            var regionAdapterMappings = new RegionAdapterMappings();
+            var container = ContainerHelper.CreateContainerExtension();
+            var regionBehaviorFactory = new RegionBehaviorFactory(container);
+
+            container.RegisterInstance<IContainerExtension>(container);
+            container.RegisterInstance<IModuleCatalog>(new ModuleCatalog());
+            container.RegisterInstance<IModuleInitializer>(mockedModuleInitializer.Object);
+            container.RegisterInstance<IModuleManager>(mockedModuleManager.Object);
+            container.RegisterInstance<RegionAdapterMappings>(regionAdapterMappings);
+            
+            container.RegisterSingleton(typeof(RegionAdapterMappings), typeof(RegionAdapterMappings));
+            container.RegisterSingleton(typeof(IRegionManager), typeof(RegionManager));
+            container.RegisterSingleton(typeof(IEventAggregator), typeof(EventAggregator));
+            container.RegisterSingleton(typeof(IRegionViewRegistry), typeof(RegionViewRegistry));
+            container.RegisterSingleton(typeof(IRegionBehaviorFactory), typeof(RegionBehaviorFactory));
+            container.RegisterSingleton(typeof(IRegionNavigationJournalEntry), typeof(RegionNavigationJournalEntry));
+            container.RegisterSingleton(typeof(IRegionNavigationJournal), typeof(RegionNavigationJournal));
+            container.RegisterSingleton(typeof(IRegionNavigationService), typeof(RegionNavigationService));
+            container.RegisterSingleton(typeof(IRegionNavigationContentLoader), typeof(RegionNavigationContentLoader));
+
+            container.RegisterInstance<SelectorRegionAdapter>(new SelectorRegionAdapter(regionBehaviorFactory));
+            container.RegisterInstance<ItemsControlRegionAdapter>(new ItemsControlRegionAdapter(regionBehaviorFactory));
+            container.RegisterInstance<ContentControlRegionAdapter>(new ContentControlRegionAdapter(regionBehaviorFactory));
+
+            var bootstrapper = new MockedContainerBootstrapper(container.GetBaseContainer());
+            bootstrapper.Run(false);
+
+            mockedModuleManager.Verify(mm => mm.Run(), Times.Once());
+        }
+
+        [StaFact]
+        public void RunShouldCallTheMethodsInOrder()
+        {
+            var bootstrapper = new MockBootstrapper();
+            bootstrapper.Run();
+
+            Assert.Equal("CreateLogger", bootstrapper.MethodCalls[0]);
+            Assert.Equal("CreateModuleCatalog", bootstrapper.MethodCalls[1]);
+            Assert.Equal("ConfigureModuleCatalog", bootstrapper.MethodCalls[2]);
+            Assert.Equal("CreateContainer", bootstrapper.MethodCalls[3]);
+            Assert.Equal("ConfigureContainer", bootstrapper.MethodCalls[4]);
+            Assert.Equal("ConfigureViewModelLocator", bootstrapper.MethodCalls[5]);
+            Assert.Equal("ConfigureRegionAdapterMappings", bootstrapper.MethodCalls[6]);
+            Assert.Equal("ConfigureDefaultRegionBehaviors", bootstrapper.MethodCalls[7]);
+            Assert.Equal("RegisterFrameworkExceptionTypes", bootstrapper.MethodCalls[8]);
+            Assert.Equal("CreateShell", bootstrapper.MethodCalls[9]);
+            Assert.Equal("InitializeShell", bootstrapper.MethodCalls[10]);
+            Assert.Equal("InitializeModules", bootstrapper.MethodCalls[11]);
+        }
+
+        [StaFact]
+        public void RunShouldLogBootstrapperSteps()
+        {
+            var bootstrapper = new MockBootstrapper();
+            bootstrapper.Run();
+            var messages = bootstrapper.BaseLogger.Messages;
+
+            Assert.Contains("Logger was created successfully.", messages[0]);
+            Assert.Contains("Creating module catalog.", messages[1]);
+            Assert.Contains("Configuring module catalog.", messages[2]);
+            Assert.Contains("Creating Unity container.", messages[3]);
+            Assert.Contains("Configuring the Unity container.", messages[4]);
+            Assert.Contains("Adding UnityBootstrapperExtension to container.", messages[5]);
+            Assert.Contains("Configuring the ViewModelLocator to use Unity.", messages[6]);
+            Assert.Contains("Configuring region adapters.", messages[7]);
+            Assert.Contains("Configuring default region behaviors.", messages[8]);
+            Assert.Contains("Registering Framework Exception Types.", messages[9]);
+            Assert.Contains("Creating the shell.", messages[10]);
+            Assert.Contains("Setting the RegionManager.", messages[11]);
+            Assert.Contains("Updating Regions.", messages[12]);
+            Assert.Contains("Initializing the shell.", messages[13]);
+            Assert.Contains("Initializing modules.", messages[14]);
+            Assert.Contains("Bootstrapper sequence completed.", messages[15]);
+        }
+
+        [StaFact]
+        public void RunShouldLogLoggerCreationSuccess()
+        {
+            const string expectedMessageText = "Logger was created successfully.";
+            var bootstrapper = new MockBootstrapper();
+            bootstrapper.Run();
+            var messages = bootstrapper.BaseLogger.Messages;
+
+            Assert.True(messages.Contains(expectedMessageText));
+        }
+        [StaFact]
+        public void RunShouldLogAboutModuleCatalogCreation()
+        {
+            const string expectedMessageText = "Creating module catalog.";
+            var bootstrapper = new MockBootstrapper();
+            bootstrapper.Run();
+            var messages = bootstrapper.BaseLogger.Messages;
+
+            Assert.True(messages.Contains(expectedMessageText));
+        }
+
+        [StaFact]
+        public void RunShouldLogAboutConfiguringModuleCatalog()
+        {
+            const string expectedMessageText = "Configuring module catalog.";
+            var bootstrapper = new MockBootstrapper();
+            bootstrapper.Run();
+            var messages = bootstrapper.BaseLogger.Messages;
+
+            Assert.True(messages.Contains(expectedMessageText));
+        }
+
+        [StaFact]
+        public void RunShouldLogAboutCreatingTheContainer()
+        {
+            const string expectedMessageText = "Creating Unity container.";
+            var bootstrapper = new MockBootstrapper();
+            bootstrapper.Run();
+            var messages = bootstrapper.BaseLogger.Messages;
+
+            Assert.True(messages.Contains(expectedMessageText));
+        }
+
+        [StaFact]
+        public void RunShouldLogAboutConfiguringContainer()
+        {
+            const string expectedMessageText = "Configuring the Unity container.";
+            var bootstrapper = new MockBootstrapper();
+            bootstrapper.Run();
+            var messages = bootstrapper.BaseLogger.Messages;
+
+            Assert.True(messages.Contains(expectedMessageText));
+        }
+
+        [StaFact]
+        public void RunShouldLogAboutConfiguringViewModelLocator()
+        {
+            const string expectedMessageText = "Configuring the ViewModelLocator to use Unity.";
+            var bootstrapper = new MockBootstrapper();
+            bootstrapper.Run();
+            var messages = bootstrapper.BaseLogger.Messages;
+
+            Assert.True(messages.Contains(expectedMessageText));
+        }
+
+        [StaFact]
+        public void RunShouldLogAboutConfiguringRegionAdapters()
+        {
+            const string expectedMessageText = "Configuring region adapters.";
+            var bootstrapper = new MockBootstrapper();
+            bootstrapper.Run();
+            var messages = bootstrapper.BaseLogger.Messages;
+
+            Assert.True(messages.Contains(expectedMessageText));
+        }
+
+        [StaFact]
+        public void RunShouldLogAboutConfiguringRegionBehaviors()
+        {
+            const string expectedMessageText = "Configuring default region behaviors.";
+            var bootstrapper = new MockBootstrapper();
+            bootstrapper.Run();
+            var messages = bootstrapper.BaseLogger.Messages;
+
+            Assert.True(messages.Contains(expectedMessageText));
+        }
+
+        [StaFact]
+        public void RunShouldLogAboutRegisteringFrameworkExceptionTypes()
+        {
+            const string expectedMessageText = "Registering Framework Exception Types.";
+            var bootstrapper = new MockBootstrapper();
+            bootstrapper.Run();
+            var messages = bootstrapper.BaseLogger.Messages;
+
+            Assert.True(messages.Contains(expectedMessageText));
+        }
+
+        [StaFact]
+        public void RunShouldLogAboutCreatingTheShell()
+        {
+            const string expectedMessageText = "Creating the shell.";
+            var bootstrapper = new MockBootstrapper();
+            bootstrapper.Run();
+            var messages = bootstrapper.BaseLogger.Messages;
+
+            Assert.True(messages.Contains(expectedMessageText));
+        }
+
+        [StaFact]
+        public void RunShouldLogAboutInitializingTheShellIfShellCreated()
+        {
+            const string expectedMessageText = "Initializing the shell.";
+            var bootstrapper = new MockBootstrapper();
+
+            bootstrapper.Run();
+            var messages = bootstrapper.BaseLogger.Messages;
+
+            Assert.True(messages.Contains(expectedMessageText));
+        }
+
+        [StaFact]
+        public void RunShouldNotLogAboutInitializingTheShellIfShellIsNotCreated()
+        {
+            const string expectedMessageText = "Initializing shell";
+            var bootstrapper = new MockBootstrapper { ShellObject = null };
+
+            bootstrapper.Run();
+            var messages = bootstrapper.BaseLogger.Messages;
+
+            Assert.False(messages.Contains(expectedMessageText));
+        }
+
+        [StaFact]
+        public void RunShouldLogAboutInitializingModules()
+        {
+            const string expectedMessageText = "Initializing modules.";
+            var bootstrapper = new MockBootstrapper();
+            bootstrapper.Run();
+            var messages = bootstrapper.BaseLogger.Messages;
+
+            Assert.True(messages.Contains(expectedMessageText));
+        }
+
+        [StaFact]
+        public void RunShouldLogAboutRunCompleting()
+        {
+            const string expectedMessageText = "Bootstrapper sequence completed.";
+            var bootstrapper = new MockBootstrapper();
+            bootstrapper.Run();
+            var messages = bootstrapper.BaseLogger.Messages;
+
+            Assert.True(messages.Contains(expectedMessageText));
+        }
+    }
+}
