+++ tests/Wpf/Prism.Container.Wpf.Shared/Prism.Container.Wpf.Shared.projitems
@@ -1,0 +1,24 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
+    <HasSharedItems>true</HasSharedItems>
+    <SharedGUID>bd42a7d6-a84d-4d27-9c28-7f6a2ec477f1</SharedGUID>
+  </PropertyGroup>
+  <PropertyGroup Label="Configuration">
+    <Import_RootNamespace>Prism.Container.Wpf.Tests</Import_RootNamespace>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)Fixtures\Bootstrapper\BootstrapperFixture.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Fixtures\Bootstrapper\BootstrapperNullLoggerFixture.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Fixtures\Bootstrapper\BootstrapperNullModuleCatalogFixture.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Fixtures\Bootstrapper\BootstrapperNullModuleManagerFixture.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Fixtures\Bootstrapper\BootstrapperRunMethodFixture.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Fixtures\Ioc\ContainerExtensionFixture.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Mocks\NullLoggerBootstrapper.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Mocks\NullModuleCatalogBootstrapper.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Mocks\NullModuleManagerBootstrapper.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Fixtures\Mvvm\ViewModelLocatorFixture.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Fixtures\Regions\RegionNavigationContentLoaderFixture.cs" />
+  </ItemGroup>
+</Project>
