+++ tests/Wpf/Prism.IocContainer.Wpf.Tests.Support/Mocks/MockService.cs
@@ -1,6 +1,3 @@
-
-
-
 namespace Prism.IocContainer.Wpf.Tests.Support.Mocks
 {
     public class MockService : IService
