+++ tests/Wpf/Prism.Unity.Wpf.Tests/BootstrapperNullContainerFixture.cs
@@ -1,0 +1,18 @@
+using System;
+using Prism.Container.Wpf.Mocks;
+using Prism.IocContainer.Wpf.Tests.Support;
+using Xunit;
+
+namespace Prism.Unity.Wpf.Tests
+{
+    public class BootstrapperNullContainerFixture : BootstrapperFixtureBase
+    {
+        [Fact]
+        public void RunThrowsWhenNullContainerCreated()
+        {
+            var bootstrapper = new NullContainerBootstrapper();
+
+            AssertExceptionThrownOnRun(bootstrapper, typeof(InvalidOperationException), "IUnityContainer");
+        }
+    }
+}
