+++ tests/Wpf/Prism.Unity.Wpf.Tests/Mocks/MockBootstrapper.cs
@@ -1,0 +1,150 @@
+﻿using System.Collections.Generic;
+using System.Reflection;
+using System.Windows;
+using System.Windows.Controls;
+using Unity;
+using Prism.IocContainer.Wpf.Tests.Support.Mocks;
+using Prism.Logging;
+using Prism.Modularity;
+using Prism.Regions;
+using Prism.Unity;
+using Prism.Ioc;
+
+namespace Prism.Container.Wpf.Mocks
+{
+    internal class MockBootstrapper : UnityBootstrapper
+    {
+        public List<string> MethodCalls = new List<string>();
+        public bool InitializeModulesCalled;
+        public bool ConfigureRegionAdapterMappingsCalled;
+        public RegionAdapterMappings DefaultRegionAdapterMappings;
+        public bool CreateLoggerCalled;
+        public bool CreateModuleCatalogCalled;
+        public bool ConfigureContainerCalled;
+        public bool CreateShellCalled;
+        public bool CreateContainerCalled;
+        public bool ConfigureModuleCatalogCalled;
+        public bool InitializeShellCalled;
+        public bool ConfigureViewModelLocatorCalled;
+        public bool ConfigureDefaultRegionBehaviorsCalled;
+        public DependencyObject ShellObject = new UserControl();
+
+        public DependencyObject BaseShell => base.Shell;
+
+        public IUnityContainer BaseContainer
+        {
+            get => base.Container;
+            set => base.Container = value;
+        }
+
+        public new IContainerExtension ContainerExtension => base.ContainerExtension;
+
+        public IContainerRegistry ContainerRegistry => base.ContainerExtension;
+
+        public MockLoggerAdapter BaseLogger => base.Logger as MockLoggerAdapter;
+
+        public IUnityContainer CallCreateContainer()
+        {
+            return this.CreateContainer();
+        }
+
+        protected override IUnityContainer CreateContainer()
+        {
+            this.MethodCalls.Add(MethodBase.GetCurrentMethod().Name);
+            this.CreateContainerCalled = true;
+            return base.CreateContainer();
+        }
+
+        protected override void ConfigureContainer()
+        {
+            this.MethodCalls.Add(MethodBase.GetCurrentMethod().Name);
+            this.ConfigureContainerCalled = true;
+            base.ConfigureContainer();
+        }
+
+        protected override ILoggerFacade CreateLogger()
+        {
+            this.MethodCalls.Add(MethodBase.GetCurrentMethod().Name);
+            this.CreateLoggerCalled = true;
+            return new MockLoggerAdapter();
+        }
+
+        protected override DependencyObject CreateShell()
+        {
+            this.MethodCalls.Add(MethodBase.GetCurrentMethod().Name);
+            this.CreateShellCalled = true;
+            return ShellObject;
+        }
+
+        //protected override void ConfigureServiceLocator()
+        //{
+        //    this.MethodCalls.Add(MethodBase.GetCurrentMethod().Name);
+        //    this.ConfigureServiceLocatorCalled = true;
+        //    base.ConfigureServiceLocator();
+        //}
+
+        protected override void ConfigureViewModelLocator()
+        {
+            this.MethodCalls.Add(MethodBase.GetCurrentMethod().Name);
+            this.ConfigureViewModelLocatorCalled = true;
+            base.ConfigureViewModelLocator();
+        }
+
+        protected override IModuleCatalog CreateModuleCatalog()
+        {
+            this.MethodCalls.Add(MethodBase.GetCurrentMethod().Name);
+            this.CreateModuleCatalogCalled = true;
+            return base.CreateModuleCatalog();
+        }
+
+        protected override void ConfigureModuleCatalog()
+        {
+            this.MethodCalls.Add(MethodBase.GetCurrentMethod().Name);
+            this.ConfigureModuleCatalogCalled = true;
+            base.ConfigureModuleCatalog();
+        }
+
+        protected override void InitializeShell()
+        {
+            this.MethodCalls.Add(MethodBase.GetCurrentMethod().Name);
+            this.InitializeShellCalled = true;
+            // no op
+        }
+
+        protected override void InitializeModules()
+        {
+            this.MethodCalls.Add(MethodBase.GetCurrentMethod().Name);
+            this.InitializeModulesCalled = true;
+            base.InitializeModules();
+        }
+
+        protected override IRegionBehaviorFactory ConfigureDefaultRegionBehaviors()
+        {
+            this.MethodCalls.Add(MethodBase.GetCurrentMethod().Name);
+            this.ConfigureDefaultRegionBehaviorsCalled = true;
+            return base.ConfigureDefaultRegionBehaviors();
+        }
+
+        protected override RegionAdapterMappings ConfigureRegionAdapterMappings()
+        {
+            this.MethodCalls.Add(MethodBase.GetCurrentMethod().Name);
+            ConfigureRegionAdapterMappingsCalled = true;
+            var regionAdapterMappings = base.ConfigureRegionAdapterMappings();
+
+            DefaultRegionAdapterMappings = regionAdapterMappings;
+
+            return regionAdapterMappings;
+        }
+
+        protected override void RegisterFrameworkExceptionTypes()
+        {
+            this.MethodCalls.Add(MethodBase.GetCurrentMethod().Name);
+            base.RegisterFrameworkExceptionTypes();
+        }
+
+        public void CallRegisterFrameworkExceptionTypes()
+        {
+            base.RegisterFrameworkExceptionTypes();
+        }
+    }
+}
