+++ tests/Wpf/Prism.Unity.Wpf.Tests/Mocks/MockedContainerBootstrapper.cs
@@ -1,0 +1,39 @@
+﻿using System.Windows;
+using System.Windows.Controls;
+using Prism.Logging;
+using Prism.Unity;
+using Unity;
+
+namespace Prism.Container.Wpf.Mocks
+{
+    internal class MockedContainerBootstrapper : UnityBootstrapper
+    {
+        private readonly IUnityContainer container;
+        public ILoggerFacade BaseLogger => base.Logger;
+
+        public void CallConfigureContainer()
+        {
+            base.ConfigureContainer();
+        }
+
+        public MockedContainerBootstrapper(IUnityContainer container)
+        {
+            this.container = container;
+        }
+
+        protected override IUnityContainer CreateContainer()
+        {
+            return container;
+        }
+
+        protected override DependencyObject CreateShell()
+        {
+            return new UserControl();
+        }
+
+        protected override void InitializeShell()
+        {
+            // no op
+        }
+    }
+}
