+++ tests/Wpf/Prism.Unity.Wpf.Tests/Mocks/NullLoggerBootstrapper.cs
@@ -1,0 +1,8 @@
+﻿using Prism.Unity;
+
+namespace Prism.Container.Wpf.Mocks
+{
+    internal partial class NullLoggerBootstrapper : UnityBootstrapper
+    {
+    }
+}
