--- tests/Wpf/Prism.Unity.Wpf.Tests/UnityBootstrapperNullLoggerFixture.cs
@@ -1,40 +1,0 @@
-
-
-using System;
-using System.Windows;
-using Xunit;
-using Prism.IocContainer.Wpf.Tests.Support;
-using Prism.Logging;
-
-namespace Prism.Unity.Wpf.Tests
-{
-    
-    public class UnityBootstrapperNullLoggerFixture : BootstrapperFixtureBase
-    {
-        [Fact]
-        public void NullLoggerThrows()
-        {
-            var bootstrapper = new NullLoggerBootstrapper();
-
-            AssertExceptionThrownOnRun(bootstrapper, typeof(InvalidOperationException), "ILoggerFacade");
-        }
-
-        internal class NullLoggerBootstrapper : UnityBootstrapper
-        {
-            protected override ILoggerFacade CreateLogger()
-            {
-                return null;
-            }
-
-            protected override DependencyObject CreateShell()
-            {
-                throw new NotImplementedException();
-            }
-
-            protected override void InitializeShell()
-            {
-                throw new NotImplementedException();
-            }
-        }
-    }
-}
