--- tests/Wpf/Prism.Unity.Wpf.Tests/UnityBootstrapperNullModuleCatalogFixture.cs
@@ -1,41 +1,0 @@
-
-
-using System;
-using System.Windows;
-using Xunit;
-using Prism.IocContainer.Wpf.Tests.Support;
-using Prism.Modularity;
-
-namespace Prism.Unity.Wpf.Tests
-{
-    
-    public class UnityBootstrapperNullModuleCatalogFixture : BootstrapperFixtureBase
-    {
-        [Fact]
-        public void NullModuleCatalogThrowsOnDefaultModuleInitialization()
-        {
-            var bootstrapper = new NullModuleCatalogBootstrapper();
-
-            AssertExceptionThrownOnRun(bootstrapper, typeof(InvalidOperationException), "IModuleCatalog");
-        }
-
-        private class NullModuleCatalogBootstrapper : UnityBootstrapper
-        {
-            protected override IModuleCatalog CreateModuleCatalog()
-            {
-                return null;
-            }
-
-            protected override DependencyObject CreateShell()
-            {
-                throw new NotImplementedException();
-            }
-
-            protected override void InitializeShell()
-            {
-                throw new NotImplementedException();
-            }
-        }
-        
-    }
-}
