--- tests/Wpf/Prism.Unity.Wpf.Tests/UnityBootstrapperNullModuleManagerFixture.cs
@@ -1,56 +1,0 @@
-using System.Windows;
-using Unity;
-using Xunit;
-using Prism.Logging;
-using Prism.Regions;
-
-namespace Prism.Unity.Wpf.Tests
-{
-    
-    public class UnityBootstrapperNullModuleManagerFixture
-    {
-        [Fact]
-        public void RunShouldNotCallInitializeModulesWhenModuleManagerNotFound()
-        {
-            var bootstrapper = new NullModuleManagerBootstrapper();
-
-            bootstrapper.Run();
-
-            Assert.False(bootstrapper.InitializeModulesCalled);
-        }
-
-        private class NullModuleManagerBootstrapper : UnityBootstrapper
-        {
-            public bool InitializeModulesCalled;
-
-            protected override void ConfigureContainer()
-            {
-                //base.RegisterDefaultTypesIfMissing();
-
-                Container.RegisterInstance<ILoggerFacade>(Logger);
-
-                this.Container.RegisterInstance(this.ModuleCatalog);
-            }
-
-            protected override IRegionBehaviorFactory ConfigureDefaultRegionBehaviors()
-            {
-                return null;
-            }
-
-            protected override RegionAdapterMappings ConfigureRegionAdapterMappings()
-            {
-                return null;
-            }
-
-            protected override DependencyObject CreateShell()
-            {
-                return null;
-            }
-
-            protected override void InitializeModules()
-            {
-                this.InitializeModulesCalled = true;
-            }
-        }
-    }
-}
